Re-use squelching information for future updates to a ticket in same request
Prior to this, information about squelching was only associated with the
initial correspond/comment, and not later changes (such as On Owner
Change), nor TransactionBatch scrips.  When updating tickets via the
web, store the squelching details on the the ticket object, and re-use
them for the lifetime of the object -- that is, the checkboxes now seen
in PreviewScrips for TransactionBatch changes are now effective.