\Q and \E only affect within their given string
Splitting the \Q and \E into differing double-quoted strings has no
effect -- and as $domain is not substituted into either of them, it does
not gain the beneficial effects.  That is, if $domain is "example.com"
the generated string will contain "(?:foo|bar)\@example.com" without the
"." escaped.

Embed the $domain substitution between the \Q and \E so it is effective.
This was detected by a warning found in the previous commit.